+++ tests/integration/element__get__query__frame_context__with_test.py
@@ -42,7 +42,7 @@ def test_actions_within_frame_context(session_browser):
 
         # THEN
         text_area.element('p').should(
-            have.js_property('innerHTML').value(
+            have.property_('innerHTML').value(
                 'Hello, World!',
             )
         )
@@ -60,7 +60,7 @@ def test_actions_within_frame_context(session_browser):
 
         # THEN
         text_area.element('p').should(
-            have.js_property('innerHTML').value('<strong>Hello, World!</strong>')
+            have.property_('innerHTML').value('<strong>Hello, World!</strong>')
         )
 
         # WHEN (just one more example)
@@ -70,7 +70,7 @@ def test_actions_within_frame_context(session_browser):
 
         # THEN
         text_area.element('p').should(
-            have.js_property('innerHTML').value(
+            have.property_('innerHTML').value(
                 '<strong>New content</strong>',
             )
         )